refactor(adr-018): fix HIGH review findings — move LlmReviewFinding to types.ts, remove wrong re-export
- Move LlmReviewFinding from semantic-review.ts to src/operations/types.ts (shared neutral types file)
  so adversarial-review.ts imports from ./types instead of cross-sibling ./semantic-review
- Remove AdversarialReviewConfig re-export from semantic-review.ts (wrong file of ownership;
  adversarial-review.ts already imports it directly from ../review/types)
- Export LlmReviewFinding from src/operations index barrel
